Page Title Update
Using the include function tag to bring in a different template file,
you may not know which page you are on from that included file, so this
allows you to display the page title in that file.

An example of this could be a header file with the title tag that gets
displayed on all pages in the template.

+ Added title function tag
+ Added title variable to routes in main config file